backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-generativeai==0.3.2
librosa==0.10.1
soundfile==0.12.1
python-multipart==0.0.9
websockets==12.0
